refactor:added orgdetail hook,organization.ts
